Simplify Event/AbstractEvent/MultiDayEventInterface etc monstrosity
