+++ src/Acts/DiaryBundle/Event/EventInterface.php
@@ -4,9 +4,6 @@ namespace Acts\DiaryBundle\Event;
 
 /**
  * Class EventInterface
- *
- * Interface which SingleDayEventInterface and MultiDayEventInterface extend. This shouldn't be used directly - use
- * one of the child classes instead.
  */
 interface EventInterface
 {
@@ -23,6 +20,20 @@ interface EventInterface
      * @return string
      */
     public function getVenue();
+
+    /**
+     * The first date on which the event takes place
+     *
+     * @return \DateTime
+     */
+    public function getStartDate();
+
+    /**
+     * The last date on which the event takes place
+     *
+     * @return \DateTime
+     */
+    public function getEndDate();
 
     /**
      * The start time of the event. It is represented by a DateTime object, but only the time component is ever used
