--- src/Acts/DiaryBundle/Event/MultiDayEvent.php
@@ -1,52 +1,0 @@
-<?php
-
-namespace Acts\DiaryBundle\Event;
-
-class MultiDayEvent extends AbstractEvent implements MultiDayEventInterface
-{
-    /**
-     * @var \DateTime
-     */
-    private $start_date;
-
-    /**
-     * @var \DateTime
-     */
-    private $end_date;
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getStartDate()
-    {
-        return $this->start_date;
-    }
-
-    /**
-     * Set the first date on which the event takes place
-     *
-     * @param \DateTime $start_date
-     */
-    public function setStartDate(\DateTime $start_date)
-    {
-        $this->start_date = $start_date;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getEndDate()
-    {
-        return $this->end_date;
-    }
-
-    /**
-     * Set the last date on which the event takes place
-     *
-     * @param \DateTime $end_date
-     */
-    public function setEndDate(\DateTime $end_date)
-    {
-        $this->end_date = $end_date;
-    }
-}
